skeleton code from github

+++ main.py
@@ -1,9 +1,3 @@
-'''
-    main.py
-    AllSwap Backend Development Team
-
-    Code for backend routes. 
-'''
 from flask import *
 import sqlite3, hashlib, os
 from werkzeug.utils import secure_filename
@@ -14,46 +8,33 @@ UPLOAD_FOLDER = 'static/uploads'
 ALLOWED_EXTENSIONS = set(['jpeg', 'jpg', 'png', 'gif'])
 app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 
-
-def getLogin():
-    with sqlite3.connect('database.db') as db:
-        cur = db.cursor()
+def getLoginDetails():
+    with sqlite3.connect('database.db') as conn:
+        cur = conn.cursor()
         if 'email' not in session:
             loggedIn = False
             firstName = ''
             noOfItems = 0
         else:
             loggedIn = True
-            cur.execute("SELECT userId, firstName FROM users WHERE email = ?",
-                        (session['email'], ))
+            cur.execute("SELECT userId, firstName FROM users WHERE email = ?", (session['email'], ))
             userId, firstName = cur.fetchone()
-            cur.execute("SELECT count(productId) FROM kart WHERE userId = ?",
-                        (userId, ))
+            cur.execute("SELECT count(productId) FROM kart WHERE userId = ?", (userId, ))
             noOfItems = cur.fetchone()[0]
-    db.close()
+    conn.close()
     return (loggedIn, firstName, noOfItems)
 
-
 @app.route("/")
-def home():
-    loggedIn, firstName, noOfItems = getLogin()
-    with sqlite3.connect('database.db') as db:
-        cur = db.cursor()
-        cur.execute(
-            'SELECT productId, name, price, description, image, stock FROM products'
-        )
+def root():
+    loggedIn, firstName, noOfItems = getLoginDetails()
+    with sqlite3.connect('database.db') as conn:
+        cur = conn.cursor()
+        cur.execute('SELECT productId, name, price, description, image, stock FROM products')
         itemData = cur.fetchall()
         cur.execute('SELECT categoryId, name FROM categories')
         categoryData = cur.fetchall()
-    itemData = parse(itemData)
-    return render_template(
-        'home.html',
-        itemData=itemData,
-        loggedIn=loggedIn,
-        firstName=firstName,
-        noOfItems=noOfItems,
-        categoryData=categoryData)
-
+    itemData = parse(itemData)   
+    return render_template('home.html', itemData=itemData, loggedIn=loggedIn, firstName=firstName, noOfItems=noOfItems, categoryData=categoryData)
 
 @app.route("/add")
 def admin():
@@ -63,7 +44,6 @@ def admin():
         categories = cur.fetchall()
     conn.close()
     return render_template('add.html', categories=categories)
-
 
 @app.route("/addItem", methods=["GET", "POST"])
 def addItem():
@@ -83,30 +63,24 @@ def addItem():
         with sqlite3.connect('database.db') as conn:
             try:
                 cur = conn.cursor()
-                cur.execute(
-                    '''INSERT INTO products (name, price, description, image, stock, categoryId) VALUES (?, ?, ?, ?, ?, ?)''',
-                    (name, price, description, imagename, stock, categoryId))
+                cur.execute('''INSERT INTO products (name, price, description, image, stock, categoryId) VALUES (?, ?, ?, ?, ?, ?)''', (name, price, description, imagename, stock, categoryId))
                 conn.commit()
-                msg = "added successfully"
+                msg="added successfully"
             except:
-                msg = "error occured"
+                msg="error occured"
                 conn.rollback()
         conn.close()
         print(msg)
         return redirect(url_for('root'))
 
-
 @app.route("/remove")
 def remove():
     with sqlite3.connect('database.db') as conn:
         cur = conn.cursor()
-        cur.execute(
-            'SELECT productId, name, price, description, image, stock FROM products'
-        )
+        cur.execute('SELECT productId, name, price, description, image, stock FROM products')
         data = cur.fetchall()
     conn.close()
     return render_template('remove.html', data=data)
-
 
 @app.route("/removeItem")
 def removeItem():
@@ -114,8 +88,7 @@ def removeItem():
     with sqlite3.connect('database.db') as conn:
         try:
             cur = conn.cursor()
-            cur.execute('DELETE FROM products WHERE productID = ?',
-                        (productId, ))
+            cur.execute('DELETE FROM products WHERE productID = ?', (productId, ))
             conn.commit()
             msg = "Deleted successsfully"
         except:
@@ -125,40 +98,25 @@ def removeItem():
     print(msg)
     return redirect(url_for('root'))
 
-
 @app.route("/displayCategory")
 def displayCategory():
-    loggedIn, firstName, noOfItems = getLoginDetails()
-    categoryId = request.args.get("categoryId")
-    with sqlite3.connect('database.db') as conn:
-        cur = conn.cursor()
-        cur.execute(
-            "SELECT products.productId, products.name, products.price, products.image, categories.name FROM products, categories WHERE products.categoryId = categories.categoryId AND categories.categoryId = ?",
-            (categoryId, ))
-        data = cur.fetchall()
-    conn.close()
-    categoryName = data[0][4]
-    data = parse(data)
-    return render_template(
-        'displayCategory.html',
-        data=data,
-        loggedIn=loggedIn,
-        firstName=firstName,
-        noOfItems=noOfItems,
-        categoryName=categoryName)
-
+        loggedIn, firstName, noOfItems = getLoginDetails()
+        categoryId = request.args.get("categoryId")
+        with sqlite3.connect('database.db') as conn:
+            cur = conn.cursor()
+            cur.execute("SELECT products.productId, products.name, products.price, products.image, categories.name FROM products, categories WHERE products.categoryId = categories.categoryId AND categories.categoryId = ?", (categoryId, ))
+            data = cur.fetchall()
+        conn.close()
+        categoryName = data[0][4]
+        data = parse(data)
+        return render_template('displayCategory.html', data=data, loggedIn=loggedIn, firstName=firstName, noOfItems=noOfItems, categoryName=categoryName)
 
 @app.route("/account/profile")
 def profileHome():
     if 'email' not in session:
         return redirect(url_for('root'))
     loggedIn, firstName, noOfItems = getLoginDetails()
-    return render_template(
-        "profileHome.html",
-        loggedIn=loggedIn,
-        firstName=firstName,
-        noOfItems=noOfItems)
-
+    return render_template("profileHome.html", loggedIn=loggedIn, firstName=firstName, noOfItems=noOfItems)
 
 @app.route("/account/profile/edit")
 def editProfile():
@@ -167,18 +125,10 @@ def editProfile():
     loggedIn, firstName, noOfItems = getLoginDetails()
     with sqlite3.connect('database.db') as conn:
         cur = conn.cursor()
-        cur.execute(
-            "SELECT userId, email, firstName, lastName, address1, address2, zipcode, city, state, country, phone FROM users WHERE email = ?",
-            (session['email'], ))
+        cur.execute("SELECT userId, email, firstName, lastName, address1, address2, zipcode, city, state, country, phone FROM users WHERE email = ?", (session['email'], ))
         profileData = cur.fetchone()
     conn.close()
-    return render_template(
-        "editProfile.html",
-        profileData=profileData,
-        loggedIn=loggedIn,
-        firstName=firstName,
-        noOfItems=noOfItems)
-
+    return render_template("editProfile.html", profileData=profileData, loggedIn=loggedIn, firstName=firstName, noOfItems=noOfItems)
 
 @app.route("/account/profile/changePassword", methods=["GET", "POST"])
 def changePassword():
@@ -191,16 +141,13 @@ def changePassword():
         newPassword = hashlib.md5(newPassword.encode()).hexdigest()
         with sqlite3.connect('database.db') as conn:
             cur = conn.cursor()
-            cur.execute("SELECT userId, password FROM users WHERE email = ?",
-                        (session['email'], ))
+            cur.execute("SELECT userId, password FROM users WHERE email = ?", (session['email'], ))
             userId, password = cur.fetchone()
             if (password == oldPassword):
                 try:
-                    cur.execute(
-                        "UPDATE users SET password = ? WHERE userId = ?",
-                        (newPassword, userId))
+                    cur.execute("UPDATE users SET password = ? WHERE userId = ?", (newPassword, userId))
                     conn.commit()
-                    msg = "Changed successfully"
+                    msg="Changed successfully"
                 except:
                     conn.rollback()
                     msg = "Failed"
@@ -211,7 +158,6 @@ def changePassword():
         return render_template("changePassword.html", msg=msg)
     else:
         return render_template("changePassword.html")
-
 
 @app.route("/updateProfile", methods=["GET", "POST"])
 def updateProfile():
@@ -227,21 +173,17 @@ def updateProfile():
         country = request.form['country']
         phone = request.form['phone']
         with sqlite3.connect('database.db') as con:
-            try:
-                cur = con.cursor()
-                cur.execute(
-                    'UPDATE users SET firstName = ?, lastName = ?, address1 = ?, address2 = ?, zipcode = ?, city = ?, state = ?, country = ?, phone = ? WHERE email = ?',
-                    (firstName, lastName, address1, address2, zipcode, city,
-                     state, country, phone, email))
+                try:
+                    cur = con.cursor()
+                    cur.execute('UPDATE users SET firstName = ?, lastName = ?, address1 = ?, address2 = ?, zipcode = ?, city = ?, state = ?, country = ?, phone = ? WHERE email = ?', (firstName, lastName, address1, address2, zipcode, city, state, country, phone, email))
 
-                con.commit()
-                msg = "Saved Successfully"
-            except:
-                con.rollback()
-                msg = "Error occured"
+                    con.commit()
+                    msg = "Saved Successfully"
+                except:
+                    con.rollback()
+                    msg = "Error occured"
         con.close()
         return redirect(url_for('editProfile'))
-
 
 @app.route("/loginForm")
 def loginForm():
@@ -250,8 +192,7 @@ def loginForm():
     else:
         return render_template('login.html', error='')
 
-
-@app.route("/login", methods=['POST', 'GET'])
+@app.route("/login", methods = ['POST', 'GET'])
 def login():
     if request.method == 'POST':
         email = request.form['email']
@@ -263,25 +204,16 @@ def login():
             error = 'Invalid UserId / Password'
             return render_template('login.html', error=error)
 
-
 @app.route("/productDescription")
 def productDescription():
     loggedIn, firstName, noOfItems = getLoginDetails()
     productId = request.args.get('productId')
     with sqlite3.connect('database.db') as conn:
         cur = conn.cursor()
-        cur.execute(
-            'SELECT productId, name, price, description, image, stock FROM products WHERE productId = ?',
-            (productId, ))
+        cur.execute('SELECT productId, name, price, description, image, stock FROM products WHERE productId = ?', (productId, ))
         productData = cur.fetchone()
     conn.close()
-    return render_template(
-        "productDescription.html",
-        data=productData,
-        loggedIn=loggedIn,
-        firstName=firstName,
-        noOfItems=noOfItems)
-
+    return render_template("productDescription.html", data=productData, loggedIn = loggedIn, firstName = firstName, noOfItems = noOfItems)
 
 @app.route("/addToCart")
 def addToCart():
@@ -291,13 +223,10 @@ def addToCart():
         productId = int(request.args.get('productId'))
         with sqlite3.connect('database.db') as conn:
             cur = conn.cursor()
-            cur.execute("SELECT userId FROM users WHERE email = ?",
-                        (session['email'], ))
+            cur.execute("SELECT userId FROM users WHERE email = ?", (session['email'], ))
             userId = cur.fetchone()[0]
             try:
-                cur.execute(
-                    "INSERT INTO kart (userId, productId) VALUES (?, ?)",
-                    (userId, productId))
+                cur.execute("INSERT INTO kart (userId, productId) VALUES (?, ?)", (userId, productId))
                 conn.commit()
                 msg = "Added successfully"
             except:
@@ -305,7 +234,6 @@ def addToCart():
                 msg = "Error occured"
         conn.close()
         return redirect(url_for('root'))
-
 
 @app.route("/cart")
 def cart():
@@ -317,21 +245,12 @@ def cart():
         cur = conn.cursor()
         cur.execute("SELECT userId FROM users WHERE email = ?", (email, ))
         userId = cur.fetchone()[0]
-        cur.execute(
-            "SELECT products.productId, products.name, products.price, products.image FROM products, kart WHERE products.productId = kart.productId AND kart.userId = ?",
-            (userId, ))
+        cur.execute("SELECT products.productId, products.name, products.price, products.image FROM products, kart WHERE products.productId = kart.productId AND kart.userId = ?", (userId, ))
         products = cur.fetchall()
     totalPrice = 0
     for row in products:
         totalPrice += row[2]
-    return render_template(
-        "cart.html",
-        products=products,
-        totalPrice=totalPrice,
-        loggedIn=loggedIn,
-        firstName=firstName,
-        noOfItems=noOfItems)
-
+    return render_template("cart.html", products = products, totalPrice=totalPrice, loggedIn=loggedIn, firstName=firstName, noOfItems=noOfItems)
 
 @app.route("/removeFromCart")
 def removeFromCart():
@@ -344,8 +263,7 @@ def removeFromCart():
         cur.execute("SELECT userId FROM users WHERE email = ?", (email, ))
         userId = cur.fetchone()[0]
         try:
-            cur.execute("DELETE FROM kart WHERE userId = ? AND productId = ?",
-                        (userId, productId))
+            cur.execute("DELETE FROM kart WHERE userId = ? AND productId = ?", (userId, productId))
             conn.commit()
             msg = "removed successfully"
         except:
@@ -354,12 +272,10 @@ def removeFromCart():
     conn.close()
     return redirect(url_for('root'))
 
-
 @app.route("/logout")
 def logout():
     session.pop('email', None)
     return redirect(url_for('root'))
-
 
 def is_valid(email, password):
     con = sqlite3.connect('database.db')
@@ -367,16 +283,14 @@ def is_valid(email, password):
     cur.execute('SELECT email, password FROM users')
     data = cur.fetchall()
     for row in data:
-        if row[0] == email and row[1] == hashlib.md5(
-                password.encode()).hexdigest():
+        if row[0] == email and row[1] == hashlib.md5(password.encode()).hexdigest():
             return True
     return False
 
-
-@app.route("/register", methods=['GET', 'POST'])
+@app.route("/register", methods = ['GET', 'POST'])
 def register():
     if request.method == 'POST':
-        #Parse form data
+        #Parse form data    
         password = request.form['password']
         email = request.form['email']
         firstName = request.form['firstName']
@@ -392,11 +306,7 @@ def register():
         with sqlite3.connect('database.db') as con:
             try:
                 cur = con.cursor()
-                cur.execute(
-                    'INSERT INTO users (password, email, firstName, lastName, address1, address2, zipcode, city, state, country, phone) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)',
-                    (hashlib.md5(password.encode()).hexdigest(), email,
-                     firstName, lastName, address1, address2, zipcode, city,
-                     state, country, phone))
+                cur.execute('INSERT INTO users (password, email, firstName, lastName, address1, address2, zipcode, city, state, country, phone) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)', (hashlib.md5(password.encode()).hexdigest(), email, firstName, lastName, address1, address2, zipcode, city, state, country, phone))
 
                 con.commit()
 
@@ -407,16 +317,13 @@ def register():
         con.close()
         return render_template("login.html", error=msg)
 
-
 @app.route("/registerationForm")
 def registrationForm():
     return render_template("register.html")
 
-
 def allowed_file(filename):
     return '.' in filename and \
             filename.rsplit('.', 1)[1] in ALLOWED_EXTENSIONS
-
 
 def parse(data):
     ans = []
@@ -430,7 +337,6 @@ def parse(data):
             i += 1
         ans.append(curr)
     return ans
-
 
 if __name__ == '__main__':
     app.run(debug=True)
